+++ src/imio/smartweb/core/contents/sections/text/views.py
@@ -8,8 +8,4 @@ class TextView(SectionView):
     """Gallery Section view"""
 
     def get_scale_url(self, item):
-        request = self.request
-        scale = self.context.image_size
-        if self.context.image._height > self.context.image._width:
-            return get_scale_url(item, request, "image", scale, "portrait")
-        return get_scale_url(item, request, "image", scale, "paysage")
+        return get_scale_url(item, self.request, "image", "section_text")